requirements.txt
asgiref==3.10.0
brotli==1.2.0
click==8.3.0
colorama==0.4.6
dj-database-url==3.0.1
Django==5.2.7
django-environ==0.12.0
django-nested-admin==4.1.6
django-nested-inline==0.4.6
django-rest-knox==5.0.2
djangorestframework==3.16.1
gunicorn==23.0.0
h11==0.16.0
packaging==25.0
psycopg2-binary==2.9.11
python-monkey-business==1.1.0
sqlparse==0.5.3
tzdata==2025.2
uvicorn==0.38.0
whitenoise==6.11.0
